+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-04-27
+
+### Added
+
+- `utils/` directory with:
+  - `index.php` file to help prevent directory snooping.
+  - `getAppVersion` function to get the app version from the CHANGELOG file.
+
 ## [1.0.0] - 2024-04-27
 
 ### Added